[PANA-8059] Reorganize semantic observation mappings
